rename: SecurityUtil를 CurrentUserProvider로 이름 변경
- 클래스가 정적 데이터를 처리하지 않으므로 적절한 이름으로 변경
- 클래스의 패키지 위치를 성향에 맞게 Util이 아닌 security로 이동
